requirement.txt
albumentations
matplotlib
numpy
pandas
Pillow
tensorflow
tqdm